backend/requirements.txt
# FastAPI framework
fastapi[all]  # Includes optional dependencies for enhanced functionality
uvicorn[standard]==0.22.0  # ASGI server with standard optional dependencies

# Dependencies for working with Jupyter Notebooks
nbformat==5.9.0  # Required for handling notebook formats
nbconvert==7.8.0  # Convert Jupyter Notebooks to other formats

# Markdown support
markdown==3.5.0

# Additional dependencies
python-multipart  # For handling file uploads and form data
python-dotenv==1.0.0  # Load environment variables from a .env file
psycopg2-binary==2.9.7

# Google Cloud dependencies
google-cloud-storage==2.10.0  # Required for interacting with Google Cloud Storage

